Query CMIP6 data using ESGF search RESTful API

The Earth System Grid Federation (ESGF) is an international collaboration for the software that powers most global climate change research, notably assessments by the Intergovernmental Panel on Climate Change (IPCC).

The ESGF search service exposes RESTful APIs that can be used by clients to query the contents of the underlying search index, and return results matching the given constraints. The documentation of the APIs can be found using this link.

EsgfQuery is the workhorse for dealing with ESGF search services.

query_esgf(index_node = "https://esgf-node.ornl.gov")

Arguments

index_node

The URL to the ESGF Index Node. Default is to use the ORNL (Oak Ridge National Laboratory) Index Node. Current possible values could be:

ORNL (Oak Ridge National Laboratory), USA: https://esgf-node.ornl.gov. The default value.

LLNL (Lawrence Livermore National Laboratory), USA: https://esgf-node.llnl.gov

NCI (National Computational Infrastructure), Australia: https://esgf.nci.org.au

IPSL (Institut Pierre-Simon Laplace), France: https://esgf-node.ipsl.upmc.fr

DKRZ (Deutsches Klimarechenzentrum), Germany: https://esgf-data.dkrz.de

LIU (National Academic Infrastructure for Supercomputing), Sweden: https://esg-dn1.nsc.liu.se

CEDA (Centre for Environmental Data Analysis), UK: https://esgf.ceda.ac.uk

Note

For bridge index nodes, only predefined common facets are returned. $list_fields() can be used to get all available fields, including facets.

EsgfQuery object

query_esgf() returns an EsgfQuery object, which is an R6 object with quite a few methods that can be classified into 3 categories:

Value listing: methods to list all possible values of facets, fields, shards, and values.

Parameter getter & setter: methods to get the query parameter values or set them before sending the actual query to the ESGF search services.

Query responses: methods to collect results for the query response.

Value listing

EsgfQuery object provides the following value-listing methods to query available facets, fields, shards, and values from the ESGF index node:

EsgfQuery$list_facets(): List all available facet names. When called, a facet listing query is sent to the index node to get all available facets for the current project (default: CMIP6).

EsgfQuery$list_fields(): List all available field names. This is useful for bridge index nodes where facet listing is not available.

EsgfQuery$list_shards(): List all available shards (ESGF index nodes) that can be queried in distributed searches.

EsgfQuery$list_values(): List all available values of specific facets.

Parameter getter & setter

The ESGF search services support a lot of parameters. The EsgfQuery contains dedicated methods to set values for most of them, including:

Most common keywords: facets, offset, limit, fields, type, replica, latest, distrib and shards.

Most common facets: project, activity_id, experiment_id, source_id, variable_id, frequency, variant_label, nominal_resolution and data_node.

All methods act in a similar way:

If input is given, the corresponding parameter is set and the updated EsgfQuery object is returned.

This makes it possible to chain different parameter setters, e.g. EsgfQuery$project("CMIP6")$frequency("day")$limit(1) sets the parameter project, frequency and limit sequentially.

For parameters that want character inputs, you can put a preceding ! to negate the constraints, e.g. EsgfQuery$project(!"CMIP6") searches for all projects except for CMIP6.

If no input is given, the current parameter value is returned. For example, directly calling EsgfQuery$project() returns the current value of the project parameter. The returned value can be two types:

NULL, i.e. there is no constraint on the corresponding parameter

An EsgfQueryParam object which is essentially a list of three elements:

value: The input values

negate: Whether there is a preceding ! in the input

name: The parameter name

Despite methods for specific keywords and facets, you can specify arbitrary query parameters using EsgfQuery$params() method. For details on the usage, please see the documentation.

Query responses

The query is not sent unless related methods are called:

EsgfQuery$count(): Count the total number of records that match the query.

You can return only the total number of matched record by calling EsgfQuery$count(facets = FALSE)

You can also count the matched records for specified facets, e.g. EsgfQuery$count(facets = c("source_id", "activity_id"))

EsgfQuery$collect(): Collect the query results and format it into an EsgfQueryResultDataset object.

Bridge Index Nodes

Some ESGF index nodes are "bridge" nodes that have certain limitations compared to standard index nodes. When using a bridge index node (e.g., https://esgf-node.ornl.gov/esgf-1-5-bridge), the following restrictions apply:

The fields parameter is not supported. All available fields are always returned.

The retracted parameter is not supported and will be ignored.

Wget script generation is not supported. Calling $url(wget = TRUE) will result in an error.

Facet listing is not available. $list_facets() will return a predefined set of common facets instead. Use $list_fields() to get all available fields.

Other helpers

EsgfQuery object also provides several other helper functions:

Query URL generation:

EsgfQuery$url(): Returns the actual query URL or the wget script URL which can be used to download all files matching the given constraints.

State persistence:

EsgfQuery$save(): Save the query state to a JSON file for later use.

EsgfQuery$load(): Restore the query state from a JSON file created by $save().

Display:

EsgfQuery$print(): Print a summary of the current EsgfQuery object including the index node URL, the last query timestamp, and all query parameters.

Author

Hongyuan Jia

Methods

Public methods

EsgfQuery$new()

EsgfQuery$list_facets()

EsgfQuery$list_fields()

EsgfQuery$list_shards()

EsgfQuery$list_values()

EsgfQuery$project()

EsgfQuery$activity_id()

EsgfQuery$experiment_id()

EsgfQuery$source_id()

EsgfQuery$variable_id()

EsgfQuery$frequency()

EsgfQuery$variant_label()

EsgfQuery$nominal_resolution()

EsgfQuery$data_node()

EsgfQuery$facets()

EsgfQuery$fields()

EsgfQuery$shards()

EsgfQuery$replica()

EsgfQuery$latest()

EsgfQuery$limit()

EsgfQuery$offset()

EsgfQuery$distrib()

EsgfQuery$params()

EsgfQuery$url()

EsgfQuery$count()

EsgfQuery$collect()

EsgfQuery$save()

EsgfQuery$load()

EsgfQuery$print()

EsgfQuery$clone()

Method new()

Create a new EsgfQuery object

Usage

EsgfQuery$new(index_node = "https://esgf-node.ornl.gov")

Arguments

index_node

The URL to the ESGF Index Node. Default is to use the ORNL (Oak Ridge National Laboratory) Index Node. Current possible values could be:

ORNL (Oak Ridge National Laboratory), USA: https://esgf-node.ornl.gov. The default value.

LLNL (Lawrence Livermore National Laboratory), USA: https://esgf-node.llnl.gov

NCI (National Computational Infrastructure), Australia: https://esgf.nci.org.au

IPSL (Institut Pierre-Simon Laplace), France: https://esgf-node.ipsl.upmc.fr

DKRZ (Deutsches Klimarechenzentrum), Germany: https://esgf-data.dkrz.de

LIU (National Academic Infrastructure for Supercomputing), Sweden: https://esg-dn1.nsc.liu.se

CEDA (Centre for Environmental Data Analysis), UK: https://esgf.ceda.ac.uk

Returns

An EsgfQuery object.

Examples

\dontrun{
q <- EsgfQuery$new(index_node = "https://esgf-node.ornl.gov")
q
}

Method list_facets()

List all available facet names

Usage

EsgfQuery$list_facets(force = FALSE)

Arguments

force

By default, every facet listing query is cached and reused when possible. If TRUE, the previous cache is abandoned and a new query is re-sent and cached. Default: FALSE.

Returns

A character vector.

Examples

\dontrun{
q$list_facets()
}

Method list_fields()

List all available field names

Usage

EsgfQuery$list_fields(force = FALSE)

Arguments

force

By default, every field listing query is cached and reused when possible. If TRUE, the previous cache is abandoned and a new query is re-sent and cached. Default: FALSE.

Returns

A character vector or NULL if no facet listing is found.

Examples

\dontrun{
q$list_fields()
}

Method list_shards()

List all available shards.

Usage

EsgfQuery$list_shards(force = FALSE)

Arguments

force

By default, every shard listing query is cached and reused when possible. If TRUE, the previous cache is abandoned and a new query is re-sent and cached. Default: FALSE.

Returns

A character vector or NULL if no shard listing is found.

Examples

\dontrun{
q$list_shards()
}

Method list_values()

List all available values of specific facets.

Usage

EsgfQuery$list_values(facets, force = FALSE)

Arguments

facets

A character vector giving the facet names.

force

By default, every value listing query is cached and reused when possible. If TRUE, the previous cache is abandoned and a new query is re-sent and cached. Default: FALSE.

Returns

If length(facets) == 1, a named character vector giving the facet value counts. Otherwise, a list of named character vectors of the same length as facets.

Examples

\dontrun{
q$list_values(c("activity_id", "experiment_id"))
}

Method project()

Get or set the project facet parameter.

Usage

EsgfQuery$project(value = "CMIP6")

Arguments

value

The parameter value. Default: "CMIP6". There are two options:

If value is not given, current value is returned.

A character vector or NULL. Note that you can put a preceding ! or - to negate the facet constraints. For example, $project(!c("CMIP6")) and $project(-c("CMIP6")) search for all projects except for CMIP6.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$project()

# set the parameter
q$project("CMIP6")

# remove the parameter
q$project(NULL)
}

Method activity_id()

Get or set the activity_id facet parameter.

Usage

EsgfQuery$activity_id(value)

Arguments

value

The parameter value. Default: NULL. There are two options:

If value is not given, current value is returned.

A character vector or NULL. Note that you can put a preceding ! or - to negate the facet constraints. For example, $activity_id(!c("C4MIP", "GeoMIP")) and $activity_id(-c("C4MIP", "GeoMIP")) search for all activity_ids except for C4MIP and GeoMIP.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$activity_id()

# set the parameter
q$activity_id("ScenarioMIP")

# negate the constraints
q$activity_id(!c("CFMIP", "ScenarioMIP"))

# remove the parameter
q$activity_id(NULL)
}

Method experiment_id()

Get or set the experiment_id facet parameter.

Usage

EsgfQuery$experiment_id(value)

Arguments

value

The parameter value. Default: NULL. There are two options:

If value is not given, current value is returned.

A character vector or NULL. Note that you can put a preceding ! or - to negate the facet constraints. For example, $experiment_id(!c("ssp126", "ssp245")) and $experiment_id(-c("ssp126", "ssp245")) search for all experiment_ids except for ssp126 and ssp245.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$experiment_id()

# set the parameter
q$experiment_id(c("ssp126", "ssp585"))

# negate the constraints
q$experiment_id(!c("ssp126", "ssp585"))

# remove the parameter
q$experiment_id(NULL)
}

Method source_id()

Get or set the source_id facet parameter.

Usage

EsgfQuery$source_id(value)

Arguments

value

The parameter value. Default: NULL. There are two options:

If value is not given, current value is returned.

A character vector or NULL. Note that you can put a preceding ! or - to negate the facet constraints. For example, $source_id(!c("CESM2", "CESM2-FV2")) and $source_id(-c("CESM2", "CESM2-FV2")) search for all source_ids except for CESM2 and CESM2-FV2.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$source_id()

# set the parameter
q$source_id(c("BCC-CSM2-MR", "CESM2"))

# negate the constraints
q$source_id(!c("BCC-CSM2-MR", "CESM2"))

# remove the parameter
q$source_id(NULL)
}

Method variable_id()

Get or set the variable_id facet parameter.

Usage

EsgfQuery$variable_id(value)

Arguments

value

The parameter value. Default: NULL. There are two options:

If value is not given, current value is returned.

A character vector or NULL. Note that you can put a preceding ! or - to negate the facet constraints. For example, $variable_id(!c("tas", "pr")) and $variable_id(-c("tas", "pr")) search for all variable_ids except for tas and pr.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$variable_id()

# set the parameter
q$variable_id(c("tas", "pr"))

# negate the constraints
q$variable_id(!c("tas", "pr"))

# remove the parameter
q$variable_id(NULL)
}

Method frequency()

Get or set the frequency facet parameter.

Usage

EsgfQuery$frequency(value)

Arguments

value

The parameter value. Default: NULL. There are two options:

If value is not given, current value is returned.

A character vector or NULL. Note that you can put a preceding ! or - to negate the facet constraints. For example, $frequency(!c("day", "mon")) and $frequency(-c("day", "mon")) search for all frequencys except for day and mon.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$frequency()

# set the parameter
q$frequency(c("1hr", "day"))

# negate the constraints
q$frequency(!c("1hr", "day"))

# remove the parameter
q$frequency(NULL)
}

Method variant_label()

Get or set the variant_label facet parameter.

Usage

EsgfQuery$variant_label(value)

Arguments

value

The parameter value. Default: NULL. There are two options:

If value is not given, current value is returned.

A character vector or NULL. Note that you can put a preceding ! or - to negate the facet constraints. For example, $variant_label(!c("r1i1p1f1", "r2i1p1f1")) and $variant_label(-c("r1i1p1f1", "r2i1p1f1")) search for all variant_labels except for r1i1p1f1 and r2i1p1f1.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$variant_label()

# set the parameter
q$variant_label(c("r1i1p1f1", "r1i2p1f1"))

# negate the constraints
q$variant_label(!c("r1i1p1f1", "r1i2p1f1"))

# remove the parameter
q$variant_label(NULL)
}

Method nominal_resolution()

Get or set the nominal_resolution facet parameter.

Usage

EsgfQuery$nominal_resolution(value)

Arguments

value

The parameter value. Default: NULL. There are two options:

If value is not given, current value is returned.

A character vector or NULL. Note that you can put a preceding ! or - to negate the facet constraints. For example, $nominal_resolution(!c("50 km", "1x1 degree")) and $nominal_resolution(-c("50 km", "1x1 degree")) search for all nominal_resolutions except for 50 km and 1x1 degree.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$nominal_resolution()

# set the parameter
q$nominal_resolution(c("100 km", "1x1 degree"))

# negate the constraints
q$nominal_resolution(!c("100 km", "1x1 degree"))

# remove the parameter
q$nominal_resolution(NULL)
}

Method data_node()

Get or set the data_node parameter.

Usage

EsgfQuery$data_node(value)

Arguments

value

The parameter value. Default: NULL. There are two options:

If value is not given, current value is returned.

A character vector or NULL. Note that you can put a preceding ! or - to negate the facet constraints. For example, $data_node(!c("cmip.bcc.cma.cn", "esg.camscma.cn")) and $data_node(-c("cmip.bcc.cma.cn", "esg.camscma.cn")) search for all data_nodes except for cmip.bcc.cma.cn and esg.camscma.cn.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$data_node()

# set the parameter
q$data_node("esg.lasg.ac.cn")

# negate the constraints
q$data_node(!"esg.lasg.ac.cn")

# remove the parameter
q$data_node(NULL)
}

Method facets()

Get or set the facets parameter for facet counting query.

Note that $facets() only affects $count() method when sending a query of facet counting.

Usage

EsgfQuery$facets(value)

Arguments

value

The facet parameter value. Default: NULL. There are two options:

If value is not given, current value is returned.

A character vector or NULL. The special notation "*" can be used to indicate that all available facets should be considered.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$facets()

# set the facets
q$facets(c("activity_id", "source_id"))

# use all available facets
q$facets("*")
}

Method fields()

Get or set the fields parameter.

By default, all available metadata fields are returned for each query. $fields() can be used to limit the number of fields returned in the query response. However, the following fields are always included in the results: access, id, index_node, number_of_aggregations, number_of_files, size, url.

Usage

EsgfQuery$fields(value = "*")

Arguments

value

The facet parameter value. Default: "*". There are two options:

If value is not given, current value is returned.

A character vector or NULL. The special notation "*" can be used to indicate that all available fields should be considered.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$fields()

# set the fields
q$fields(c("activity_id", "source_id"))

# use all available fields
q$fields("*")

# remove the parameter
# act the same as above because the default `fields` in ESGF search
# services is `*` if `fields` is not specified
q$fields(NULL)
}

Method shards()

Get or set the shards parameter.

By default, a distributed query targets all ESGF Nodes. $shards() can be used to execute a distributed search that targets only one or more specific nodes.

Usage

EsgfQuery$shards(value)

Arguments

value

The facet parameter value. There are two options:

If value is not given, current value is returned.

A character vector or NULL.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$shards()

# set the parameter
q$shards("localhost:8983/solr/datasets")

# negate the constraints
q$shards(!"localhost:8983/solr/datasets")

# only applicable for distributed queries
q$distrib(FALSE)$shards("localhost:8983/solr/datasets") # Error

# remove the parameter
q$shards(NULL)
}

Method replica()

Get or set the replica parameter.

By default, a query returns all records (masters and replicas) matching the search criteria, i.e. $replica(NULL). To return only master records, use $replica(FALSE); to return only replicas, use $replica(TRUE).

Usage

EsgfQuery$replica(value)

Arguments

value

The facet parameter value. Default: NULL. There are two options:

If value is not given, current value is returned.

A flag or NULL.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$replica()

# set the parameter
q$replica(TRUE)

# remove the parameter
q$replica(NULL)
}

Method latest()

Get or set the latest parameter.

By default, a query to the ESGF search services returns only the very last, up-to-date version of the matching records, i.e. $latest(TRUE). You can use $latest(FALSE) to return all versions.

Usage

EsgfQuery$latest(value = TRUE)

Arguments

value

The facet parameter value. Default: TRUE. There are two options:

If value is not given, current value is returned.

A flag.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$latest()

# set the parameter
q$latest(TRUE)
}

Method limit()

Get or set the limit parameter.

$limit() can be used to limit the number of records to return. Note that the maximum number of records to return per query for ESGF search services is 10,000. A warning is issued if input value is greater than that. In this case, limit will be reset to 10,000.

Usage

EsgfQuery$limit(value = 10L)

Arguments

value

The facet parameter value. Default: 10. There are two options:

If value is not given, current value is returned.

An integer.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$limit()

# set the parameter
q$limit(10L)

# `limit` is reset to the allowed maximum query limit if input is greater than that
q$limit(12000L) # warning
}

Method offset()

Get or set the offset parameter.

If the query returns records that exceed the limit number, $offset() can be used to paginate through the available results.

Usage

EsgfQuery$offset(value = 0L)

Arguments

value

The facet parameter value. Default: 0. There are two options:

If value is not given, current value is returned.

An integer.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$offset()

# set the parameter
q$offset(0L)
}

Method distrib()

Get or set the distrib facet

By default, the query is sent to all ESGF Nodes, i.e. $distrib(TRUE). $distrib(FALSE) can be used to execute the query only on the target node.

Usage

EsgfQuery$distrib(value = TRUE)

Arguments

value

The facet parameter value. Default: TRUE. There are two options:

If value is not given, current value is returned.

A flag.

Returns

If value is given, the modified EsgfQuery object.

Otherwise, an EsgfQueryParam object which is essentially a list of three elements:

value: input values.

negate: Whether there is a preceding ! or -.

name: Parameter name.

Examples

\dontrun{
# get current value
q$distrib()

# set the parameter
q$distrib(TRUE)
}

Method params()

Get or set other parameters.

$params() can be used to specify other parameters that do not have a dedicated method, e.g. version, master_id, etc. It can also be used to overwrite existing parameter values specified using methods like $activity_id().

Usage

EsgfQuery$params(...)

Arguments

...

Parameter values to set. There are three options:

If not given, existing parameters that do not have a dedicated method are returned.

If NULL, all existing parameters that do not have a dedicated method are removed.

A named vector, e.g. $params(score = 1, table_id = "day") will set score to 1 and table_id to day. The ! notation can still be used to negate the constraints, e.g. $params(table_id = !c("3hr", "day")) searches for all table_id except for 3hr and day.

Returns

If parameters are specified, the modified EsgfQuery object.

Otherwise, an empty list for $params(NULL) or a list of EsgfQueryParam objects.

Examples

\dontrun{
# get current values
# default is an empty list (`list()`)
q$params()

# set the parameter
q$params(table_id = c("3hr", "day"), member_id = "00")
q$params()

# reset existing parameters
q$frequency("day")
q$params(frequency = "mon")
q$frequency() # frequency value has been changed using $params()

# negating the constraints is also supported
q$params(table_id = !c("3hr", "day"))

# use NULL to remove all parameters
q$params(NULL)$params()
}

Method url()

Get the URL of actual query or wget script

The wget script URL can be used to download a bash script that contains wget commands for downloading all files matching the query constraints. This is useful for batch downloading large amounts of data.

Usage

EsgfQuery$url(wget = FALSE)

Arguments

wget

Whether to return the URL of the wget script that can be used to download all files matching the given constraints. Default: FALSE.

Returns

A single string.

Examples

\dontrun{
q$url()

# get the wget script URL
q$url(wget = TRUE)

# You can download the wget script using the URL directly. For
# example, the code below downloads the script and save it as
# 'wget.sh' in R's temporary folder:
download.file(q$url(TRUE), file.path(tempdir(), "wget.sh"), mode = "wb")

}

Method count()

Send a query of facet counting and fetch the results

Usage

EsgfQuery$count(facets = TRUE)

Arguments

facets

NULL, a flag or a character vector. There are three options:

If NULL or FALSE, only the total number of matched records is returned.

If TRUE, the value of $facets() is used to limit the facets. If $facets() returns NULL, only the total count is returned. This is the default value.

If a character vector, it is used to limit the facets.

Returns

If facets equals NULL or FALSE, or $facets() returns NULL, an integer.

Otherwise, a named list with the first element always being total which is the total number of matched records. Other elements have the same length as input facets and are all named integer vectors.

Examples

\dontrun{
# get the total number of matched records
q$count(NULL) # or q$count(facets = FALSE)

# count records for specific facets
q$facets(c("activity_id", "source_id"))$count()

# same as above
q$count(facets = c("activity_id", "source_id"))
}

Method collect()

Send the actual query and fetch the results

$collect() sends the actual query with type=Dataset to the ESGF search services and returns the results as an EsgfQueryResultDataset object. The fields included depend on fields parameter. However, the following fields are always included in the results: access, id, index_node, number_of_aggregations, number_of_files, size, url.

Usage

EsgfQuery$collect(all = FALSE, limit = TRUE, params = TRUE)

Arguments

all

Whether to collect all results despite of the value of offset. Default: FALSE.

limit

Only applicable when all is set to TRUE. Whether to respect the current value of limit when collecting all matched records. If FALSE, the allowed maximum limit number 10000 is used. It can also be a positive integer which will be used as a temporary limit per query. Default: FALSE.

params

Whether to include facet fields that have parameter constraints explicitly set using EsgfQuery$project(), EsgfQuery$activity_id(), EsgfQuery$params() and etc. in the returned fields. For example, if you set $experiment_id("ssp585"), the experiment_id field will be included in the results when params = TRUE. Default: TRUE.

Returns

An EsgfQueryResultDataset object.

Examples

\dontrun{
# by default, all fields with constrains are included in the results
query <- query_esgf()$experiment_id("ssp585")$frequency("1hr")$fields("source_id")
res1 <- query$collect()
res1$fields

# set `params` to `FALSE` to exclude them
query$collect(params = FALSE)$fields

# collect all matched records with `query$limit()` records per query
res2 <- query$collect(all = TRUE, limit = TRUE)
identical(query$count(), res2$count())

# same as above, but collect all matched records with max allowed
# record limit per query
res3 <- query$collect(all = TRUE, limit = FALSE)
identical(res2$count(), res3$count())

# same as above, but collect all matched records with specified limit
# per query
res4 <- query$collect(all = TRUE, limit = 30)
identical(res2$count(), res4$count())
}

Method save()

Save the query into a JSON file

$save() puts main data of an EsgfQuery object into a JSON file which can be loaded to restore the current state of query using EsgfQuery$load().

Usage

EsgfQuery$save(file = "query.json", pretty = TRUE)

Arguments

file

A string indicating the JSON file path to save the data to.

pretty

Whether to add indentation whitespace to JSON output. For details, please see jsonlite::toJSON(). Default: TRUE.

Returns

The full path of the output JSON file.

Examples

\dontrun{
q$save(tempfile(fileext = ".json"))
}

Method load()

Restore the query state from an JSON file

$load() reads data of an EsgfQuery object from a JSON file created using EsgfQuery$save().

Usage

EsgfQuery$load(file)

Arguments

file

A string indicating the JSON file path to read the data from.

Returns

The modified EsgfQuery object itself.

Examples

\dontrun{
f <- tempfile(fileext = "json")

q <- query_esgf()
json <- q$save(f)
q$load(f)
}

Method print()

Print a summary of the current EsgfQuery object

$print() gives the summary of current EsgfQuery object including the index node URL, the last query timestamp, and all query parameters.

Usage

EsgfQuery$print()

Returns

The EsgfQuery object itself, invisibly.

Examples

\dontrun{
q$print()
}

Method clone()

The objects of this class are cloneable with this method.

Usage

EsgfQuery$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `EsgfQuery$new`
## ------------------------------------------------

if (FALSE) { # \dontrun{
q <- EsgfQuery$new(index_node = "https://esgf-node.ornl.gov")
q
} # }

## ------------------------------------------------
## Method `EsgfQuery$list_facets`
## ------------------------------------------------

if (FALSE) { # \dontrun{
q$list_facets()
} # }

## ------------------------------------------------
## Method `EsgfQuery$list_fields`
## ------------------------------------------------

if (FALSE) { # \dontrun{
q$list_fields()
} # }

## ------------------------------------------------
## Method `EsgfQuery$list_shards`
## ------------------------------------------------

if (FALSE) { # \dontrun{
q$list_shards()
} # }

## ------------------------------------------------
## Method `EsgfQuery$list_values`
## ------------------------------------------------

if (FALSE) { # \dontrun{
q$list_values(c("activity_id", "experiment_id"))
} # }

## ------------------------------------------------
## Method `EsgfQuery$project`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$project()

# set the parameter
q$project("CMIP6")

# remove the parameter
q$project(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$activity_id`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$activity_id()

# set the parameter
q$activity_id("ScenarioMIP")

# negate the constraints
q$activity_id(!c("CFMIP", "ScenarioMIP"))

# remove the parameter
q$activity_id(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$experiment_id`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$experiment_id()

# set the parameter
q$experiment_id(c("ssp126", "ssp585"))

# negate the constraints
q$experiment_id(!c("ssp126", "ssp585"))

# remove the parameter
q$experiment_id(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$source_id`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$source_id()

# set the parameter
q$source_id(c("BCC-CSM2-MR", "CESM2"))

# negate the constraints
q$source_id(!c("BCC-CSM2-MR", "CESM2"))

# remove the parameter
q$source_id(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$variable_id`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$variable_id()

# set the parameter
q$variable_id(c("tas", "pr"))

# negate the constraints
q$variable_id(!c("tas", "pr"))

# remove the parameter
q$variable_id(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$frequency`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$frequency()

# set the parameter
q$frequency(c("1hr", "day"))

# negate the constraints
q$frequency(!c("1hr", "day"))

# remove the parameter
q$frequency(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$variant_label`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$variant_label()

# set the parameter
q$variant_label(c("r1i1p1f1", "r1i2p1f1"))

# negate the constraints
q$variant_label(!c("r1i1p1f1", "r1i2p1f1"))

# remove the parameter
q$variant_label(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$nominal_resolution`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$nominal_resolution()

# set the parameter
q$nominal_resolution(c("100 km", "1x1 degree"))

# negate the constraints
q$nominal_resolution(!c("100 km", "1x1 degree"))

# remove the parameter
q$nominal_resolution(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$data_node`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$data_node()

# set the parameter
q$data_node("esg.lasg.ac.cn")

# negate the constraints
q$data_node(!"esg.lasg.ac.cn")

# remove the parameter
q$data_node(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$facets`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$facets()

# set the facets
q$facets(c("activity_id", "source_id"))

# use all available facets
q$facets("*")
} # }

## ------------------------------------------------
## Method `EsgfQuery$fields`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$fields()

# set the fields
q$fields(c("activity_id", "source_id"))

# use all available fields
q$fields("*")

# remove the parameter
# act the same as above because the default `fields` in ESGF search
# services is `*` if `fields` is not specified
q$fields(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$shards`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$shards()

# set the parameter
q$shards("localhost:8983/solr/datasets")

# negate the constraints
q$shards(!"localhost:8983/solr/datasets")

# only applicable for distributed queries
q$distrib(FALSE)$shards("localhost:8983/solr/datasets") # Error

# remove the parameter
q$shards(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$replica`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$replica()

# set the parameter
q$replica(TRUE)

# remove the parameter
q$replica(NULL)
} # }

## ------------------------------------------------
## Method `EsgfQuery$latest`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$latest()

# set the parameter
q$latest(TRUE)
} # }

## ------------------------------------------------
## Method `EsgfQuery$limit`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$limit()

# set the parameter
q$limit(10L)

# `limit` is reset to the allowed maximum query limit if input is greater than that
q$limit(12000L) # warning
} # }

## ------------------------------------------------
## Method `EsgfQuery$offset`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$offset()

# set the parameter
q$offset(0L)
} # }

## ------------------------------------------------
## Method `EsgfQuery$distrib`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current value
q$distrib()

# set the parameter
q$distrib(TRUE)
} # }

## ------------------------------------------------
## Method `EsgfQuery$params`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get current values
# default is an empty list (`list()`)
q$params()

# set the parameter
q$params(table_id = c("3hr", "day"), member_id = "00")
q$params()

# reset existing parameters
q$frequency("day")
q$params(frequency = "mon")
q$frequency() # frequency value has been changed using $params()

# negating the constraints is also supported
q$params(table_id = !c("3hr", "day"))

# use NULL to remove all parameters
q$params(NULL)$params()
} # }

## ------------------------------------------------
## Method `EsgfQuery$url`
## ------------------------------------------------

if (FALSE) { # \dontrun{
q$url()

# get the wget script URL
q$url(wget = TRUE)

# You can download the wget script using the URL directly. For
# example, the code below downloads the script and save it as
# 'wget.sh' in R's temporary folder:
download.file(q$url(TRUE), file.path(tempdir(), "wget.sh"), mode = "wb")

} # }

## ------------------------------------------------
## Method `EsgfQuery$count`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# get the total number of matched records
q$count(NULL) # or q$count(facets = FALSE)

# count records for specific facets
q$facets(c("activity_id", "source_id"))$count()

# same as above
q$count(facets = c("activity_id", "source_id"))
} # }

## ------------------------------------------------
## Method `EsgfQuery$collect`
## ------------------------------------------------

if (FALSE) { # \dontrun{
# by default, all fields with constrains are included in the results
query <- query_esgf()$experiment_id("ssp585")$frequency("1hr")$fields("source_id")
res1 <- query$collect()
res1$fields

# set `params` to `FALSE` to exclude them
query$collect(params = FALSE)$fields

# collect all matched records with `query$limit()` records per query
res2 <- query$collect(all = TRUE, limit = TRUE)
identical(query$count(), res2$count())

# same as above, but collect all matched records with max allowed
# record limit per query
res3 <- query$collect(all = TRUE, limit = FALSE)
identical(res2$count(), res3$count())

# same as above, but collect all matched records with specified limit
# per query
res4 <- query$collect(all = TRUE, limit = 30)
identical(res2$count(), res4$count())
} # }

## ------------------------------------------------
## Method `EsgfQuery$save`
## ------------------------------------------------

if (FALSE) { # \dontrun{
q$save(tempfile(fileext = ".json"))
} # }

## ------------------------------------------------
## Method `EsgfQuery$load`
## ------------------------------------------------

if (FALSE) { # \dontrun{
f <- tempfile(fileext = "json")

q <- query_esgf()
json <- q$save(f)
q$load(f)
} # }

## ------------------------------------------------
## Method `EsgfQuery$print`
## ------------------------------------------------

if (FALSE) { # \dontrun{
q$print()
} # }