Base class for results for ESGF query

EsgfQueryResult is a base class that represents basic query results from ESGF search RESTful API. It defines common fields and methods for results from all query types, including Dataset, File and Aggregation. Results from the three types are

In general, there is no need to create an EsgfQueryResult manually.

Author

Hongyuan Jia

Active bindings

id

A character vector indicating globally unique record identifiers.

url

A list of data.table with 3 columns:

service \[character\]: The service types, e.g. OPENDAP, HTTPServer, etc.;

url \[character\]: The actual URLs;

mime_type \[character\]: The MIME types indicating the nature and format of the corresponding document of the URLs.

size

A vector of units indicating the file sizes.

fields

A character vector indicating all fields in the results.

Methods

Public methods

EsgfQueryResult$new()

EsgfQueryResult$to_data_table()

EsgfQueryResult$count()

EsgfQueryResult$save()

EsgfQueryResult$load()

EsgfQueryResult$clone()

Method new()

Create a new EsgfQueryResult object

Usage

EsgfQueryResult$new(index_node, params, response)

Arguments

index_node

The URL to the ESGF Index Node. It should be the same as the index_node for an EsgfQuery object that collects the query results.

params

A list of query parameters.

response

The result of an query response.

Returns

An EsgfQueryResult object.

Method to_data_table()

Convert the results into a data.table

Usage

EsgfQueryResult$to_data_table(fields = NULL, formatted = NULL)

Arguments

fields

A character vector indicating the fields to put into the data.table. If NULL, all fields in the query result will be used. Possible field names can be retrieved using $fields. Default: NULL.

formatted

Whether to use formatted values for special fields, including url and size. Default: FALSE.

Returns

A data.table.

Method count()

Count the number of matched records in current result

Usage

EsgfQueryResult$count()

Returns

An integer.

Method save()

Save the result into a JSON file

$save() puts main data of an EsgfQueryResult object into a JSON file which can be loaded to restore the current state of result using EsgfQueryResult$load().

Usage

EsgfQueryResult$save(file, pretty = TRUE)

Arguments

file

A string indicating the JSON file path to save the data to.

pretty

Whether to add indentation whitespace to JSON output. For details, please see jsonlite::toJSON(). Default: TRUE.

Returns

The full path of the output JSON file.

Method load()

Restore the result state from an JSON file

$load() reads data of an EsgfQueryResult object from a JSON file created using EsgfQueryResult$save().

Usage

EsgfQueryResult$load(file)

Arguments

file

A string indicating the JSON file path to read the data from.

Returns

The modified EsgfQueryResult object itself.

Method clone()

The objects of this class are cloneable with this method.

Usage

EsgfQueryResult$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.