ESGF Query results for File type

EsgfQueryResultFile is a class that represents query results for File type from ESGF search RESTful API.

In general, there is no need to create an EsgfQueryResultDataset manually. Usually, it is created by calling EsgfQueryResultDataset$collect().

Author

Hongyuan Jia

Super class

epwshiftr::EsgfQueryResult -> EsgfQueryResultFile

Active bindings

filename

A character vector indicating file names on the sever.

url_opendap

A character vector of the OPeNDAP URLs of the files.

url_download

A character vector of the download URLs of the files.

fields

A character vector indicating all fields in the results.

Methods

Public methods

EsgfQueryResultFile$to_data_table()

EsgfQueryResultFile$print()

EsgfQueryResultFile$clone()

Inherited methods

epwshiftr::EsgfQueryResult$count()

epwshiftr::EsgfQueryResult$initialize()

epwshiftr::EsgfQueryResult$load()

epwshiftr::EsgfQueryResult$save()

Method to_data_table()

Convert the results into a data.table

Usage

EsgfQueryResultFile$to_data_table(fields = NULL, formatted = FALSE)

Arguments

fields

A character vector indicating the fields to put into the data.table. If NULL, all fields in the query result will be used. Default: NULL.

formatted

Whether to use formatted values for special fields, including url and size. Default: FALSE.

Returns

A data.table.

Method print()

Print a summary of the current dataset

Usage

EsgfQueryResultFile$print(n = 10L)

Arguments

n

An integer indicating how many items to print. If NULL, all items will be printed. Default: 10L.

Returns

The EsgfQueryResultFile object itself, invisibly.

Method clone()

The objects of this class are cloneable with this method.

Usage

EsgfQueryResultFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.