Base class for results for ESGF query

EsgResult is a base class that represents basic query results from ESGF search RESTful API. It defines common fields and methods for results from all query types, including Dataset, File and Aggregation. Results from the three types are

In general, there is no need to create an EsgResult manually.

Author

Hongyuan Jia

Active bindings

id

A character vector indicating globally unique record identifiers.

url

A list of data.table with 3 columns:

service \[character\]: The service types, e.g. OPENDAP, HTTPServer, etc.;

url \[character\]: The actual URLs;

mime_type \[character\]: The MIME types indicating the nature and format of the corresponding document of the URLs.

size

A vector of units indicating the file sizes.

fields

A character vector indicating all fields in the results, preserving the order returned by the response.

time_filter

A list describing the time range used by $filter_time(), or NULL if no result-level time filter has been recorded.

Methods

Public methods

EsgResult$new()

EsgResult$to_data_table()

EsgResult$to_dt()

EsgResult$count()

EsgResult$save()

EsgResult$load()

EsgResult$clone()

Method new()

Create a new EsgResult object

Usage

EsgResult$new(index_node, params, response, context = NULL)

Arguments

index_node

The URL to the ESGF Index Node. It should be the same as the index_node for an EsgQuery object that collects the query results.

params

A list of query parameters.

response

The result of an query response.

context

Optional saved-result context. Used internally for result-level metadata such as recorded time filters.

Returns

An EsgResult object.

Method to_data_table()

Convert the results into a data.table

Usage

EsgResult$to_data_table(fields = NULL, formatted = NULL)

Arguments

fields

A non-empty character vector indicating the fields to put into the data.table. If NULL, all fields in the query result will be used. Possible field names can be retrieved using $fields. Default: NULL.

formatted

Whether to use formatted values for special fields, including url and size. Default: FALSE.

Returns

A data.table.

Method to_dt()

Alias of $to_data_table().

Usage

EsgResult$to_dt(...)

Arguments

...

Arguments passed to $to_data_table().

Returns

A data.table.

Method count()

Count the number of matched records in current result

Usage

EsgResult$count()

Returns

An integer.

Method save()

Save the result into a JSON file

$save() puts main data of an EsgResult object into a JSON file which can be loaded to restore the current state of result using EsgResult$load().

Usage

EsgResult$save(file, pretty = TRUE)

Arguments

file

A string indicating the JSON file path to save the data to.

pretty

Whether to add indentation whitespace to JSON output. For details, please see jsonlite::toJSON(). Default: TRUE.

Returns

The full path of the output JSON file.

Method load()

Restore the result state from an JSON file

$load() reads data of an EsgResult object from a JSON file created using EsgResult$save(). Saved result fields are restored as read-only active bindings. A saved file can only be loaded by the matching result type.

Usage

EsgResult$load(file)

Arguments

file

A string indicating the JSON file path to read the data from.

Returns

The modified EsgResult object itself.

Method clone()

The objects of this class are cloneable with this method.

Usage

EsgResult$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.