Get data nodes which store CMIP6 output

Get data nodes which store CMIP6 output

get_data_node(speed_test = FALSE, timeout = 3)

Arguments

speed_test

If TRUE, use pingr::ping() to perform connection speed test on each data node. A ping column is appended in returned data.table which stores each data node response in milliseconds. This feature needs pingr package already installed. Default: FALSE.

timeout

Timeout for a ping response in seconds. Default: 3.

Value

A data.table::data.table() of 2 or 3 (when speed_test is TRUE) columns:

Column	Type	Description

data_node	character	Web address of data node

status	character	Status of data node. "UP" means OK and "DOWN" means currently not available

ping	double	Data node response in milliseconds during speed test

Examples

# \donttest{
get_data_node()
#>                     data_node status
#>  1:            aims3.llnl.gov     UP
#>  2:           cmip.bcc.cma.cn     UP
#>  3: cmip.dess.tsinghua.edu.cn     UP
#>  4:           cmip.fio.org.cn     UP
#>  5:          cordexesg.dmi.dk     UP
#>  6:     crd-esgf-drc.ec.gc.ca     UP
#>  7:      data.meteo.unican.es     UP
#>  8:  dataserver.nccs.nasa.gov     UP
#>  9:    dpesgf03.nccs.nasa.gov     UP
#> 10:   esg-cccr.tropmet.res.in     UP
#> 11:        esg-dn1.nsc.liu.se     UP
#> 12:          esg-dn1.ru.ac.th     UP
#> 13:        esg-dn2.nsc.liu.se     UP
#> 14:        esg-dn3.nsc.liu.se     UP
#> 15:            esg.camscma.cn     UP
#> 16:            esg.lasg.ac.cn     UP
#> 17:          esg2.umr-cnrm.fr     UP
#> 18:     esgdata.gfdl.noaa.gov     UP
#> 19:    esgf-cnr.hpc.cineca.it     UP
#> 20:        esgf-data.ucar.edu     UP
#> 21:    esgf-data01.diasjp.net     UP
#> 22:    esgf-data02.diasjp.net     UP
#> 23:    esgf-data03.diasjp.net     UP
#> 24:    esgf-data04.diasjp.net     UP
#> 25:     esgf-data1.diasjp.net     UP
#> 26:       esgf-data1.llnl.gov     UP
#> 27:       esgf-data2.llnl.gov     UP
#> 28:   esgf-ictp.hpc.cineca.it     UP
#> 29: esgf-nimscmip6.apcc21.org     UP
#> 30:         esgf-node.cmcc.it     UP
#> 31:        esgf-node.ornl.gov     UP
#> 32:        esgf-node2.cmcc.it     UP
#> 33:           esgf.apcc21.org     UP
#> 34:               esgf.bsc.es     UP
#> 35:           esgf.ceda.ac.uk     UP
#> 36:               esgf.dwd.de     UP
#> 37:        esgf.nccs.nasa.gov     UP
#> 38:           esgf.nci.org.au     UP
#> 39:   esgf.rcec.sinica.edu.tw     UP
#> 40:             esgf1.dkrz.de     UP
#> 41:             esgf2.dkrz.de     UP
#> 42:             esgf3.dkrz.de     UP
#> 43:     noresg.nird.sigma2.no     UP
#> 44:      dist.nmlab.snu.ac.kr   DOWN
#> 45:           dmiesgf0.dmi.dk   DOWN
#> 46:       eridanus.eoc.dlr.de   DOWN
#> 47:        esg.pik-potsdam.de   DOWN
#> 48:          esg1.umr-cnrm.fr   DOWN
#> 49:          esgf-data.csc.fi   DOWN
#> 50:              esgf.anl.gov   DOWN
#> 51:             esgf.ichec.ie   DOWN
#> 52:        polaris.pknu.ac.kr   DOWN
#> 53:         vesg.ipsl.upmc.fr   DOWN
#>                     data_node status
# }